Admin profile done. Login and password are in DB

+++ views/film-single.tpl
@@ -12,8 +12,17 @@
         <div class="card__header">
           <h4 class="title-4"><?php echo $films['name'];?></h4>
           <div class="button-block">
-            <a href="edit.php?id=<?php echo $films['id'];?>" class="button button--edit">Редактировать</a>
-            <a href="index.php?action=delete&id=<?php echo $films['id'];?>" class="button button--delete">Удалить</a>
+             <?php if (isAdmin()) { ?>
+         
+                  <a href="edit.php?id=<?php echo $films['id'];?>" class="button button--edit">Редактировать</a>
+
+            <?php }
+                  if (isAdmin()) { ?>
+         
+                   <a href="index.php?action=delete&id=<?php echo $films['id'];?>" class="button button--delete">Удалить</a>
+                  
+            <?php } ?>
+
           </div>
         </div>
         <div class="badge"><?php echo $films['genre'];?></div>
